// src/models/game-statistics.ts
export interface GameStatistics { 
    t_games: number;
    t_score: number;
    t_level: number;
    t_lines: number;
    t_game_time: number;
    t_minigame_points: number;
    t_ships_destroyed: number;
    t_shots_destroyed: number;
    t_shots_fired: number;
    t_powerups_used: number;
    t_moves: number;
    t_death_abyss: number;
    t_death_spaceship: number;
    t_death_wall: number;
}